infineon-firmware-updater: fix clearing owner for recovery
When updating the firmware in recovery mode, the tpm ownership must be
cleared before the update. Otherwise, it fails. On the other hand, the
ownership is not taken by the update itself, so requesting clearing the
owner on the next boot is not needed.

BUG=chromium:728130
BUG=chromium:762934
TEST=go through tpm powerwash-based update, successfully own the device

Change-Id: Ic5ed14e5853e1480c243f69838a670c5d1560ee0
Reviewed-on: https://chromium-review.googlesource.com/654297
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>