cros-firmware: Handle downloading of shared firmware correctly
When a model does not have its own firmware we should not try to download
an image for it from BCS. It will likely not be there, and we don't need
it anyway.

Process shared firmware in the family separately at the start. Then when
we work through the models, skip models which use shared firmware.

BUG=b:65431038,chromium:741043
TEST=emerge-coral --nodeps chromeos-firmware-coral
See that we now get just the coral tarballs from distdir:
 * Coral.9899.0.0.tbz2 SHA256 SHA512 WHIRLPOOL size ;-) ...                                                              [ ok ]
 * Coral_EC.9899.0.0.tbz2 SHA256 SHA512 WHIRLPOOL size ;-) ...                                                           [ ok ]

Change-Id: Ib42478ff29b983f097dd38154fb17105c5895f69
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/655770
Reviewed-by: Jason Clinton <jclinton@chromium.org>