Revert cryptohome sandboxing changes.
This reverts the CLs to sandbox cryptohome which cause dregressions.

This reverts commit f7ff65bd0e20ce532d97cc511e1c0ff1749ae91d
Author: Matthew Denton <mpdenton@google.com>
Date: Thu Jul 20 07:36:56 2017

Update cryptohome ebuild to create cryptohome user

Updated the cryptohome ebuild file with a pkg_preinst to create the
new user and group "cryptohome".

This reverts commit 3bb107b6d68d92f23927b99643934be3554a6668
Author: Matthew Denton <mpdenton@google.com>
Date: Fri Aug 04 22:39:04 2017

upstart: Create dircrypto keyring with owner "cryptohome"

This CL creates the dircrypto keyring with owner user and group as
"cryptohome". This is necessary in order to run cryptohome as non-root
user "cryptohome", as cryptohome needs to create keys in the dircrypto
keyring. The keyring was originally owned (and could only be modified
by) root.

BUG=chromium:741786,chromium:764540
TEST=pre-cq

Change-Id: I24a9fdd6c6251001fb807bbd5cf3674dbafca3e0
Reviewed-on: https://chromium-review.googlesource.com/665345
Commit-Ready: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>