busybox: turn off static linking
We turned this on because it was historically enabled, and because
there was a latent bug in the hush wrapper patch.  Those should be
fixed by the upstream Gentoo 5c7ecf36f0bbbe18b513d7afb82b0f7bf3428
(which we included in CL:489122), so we can drop this again.

BUG=chromium:712564,chromium:710074,chromium:764753
TEST=precq passes
TEST=recovery image still works

Change-Id: Ia8de426f7da212ef2970340adfca691d9ebfc553
Reviewed-on: https://chromium-review.googlesource.com/668036
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>