glibc: Call cros_use_gcc inside eblit-src_configure-pre.
glibc needs to be built with gcc. It also requires
that TARGET_CC is also set to gcc. To ensure that, call
cros_use_gcc in eblit-src_configure-pre.
Alternative to modifying the build is to create config/env/cross-*/glibc files
just to create a call to cros-use_gcc which will be a mess to manage.

BUG=chromium:748476
TEST=chromiumos-sdk passed.

Change-Id: I80dc95266b7bc01c7c43bfa450092b1aae1deaf8
Reviewed-on: https://chromium-review.googlesource.com/669925
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Trybot-Ready: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>