Rename model to build_target for clarity
Based on feedback in crrev.com/c/670042, it's currently unclear that we
don't have to build firmware for every model.  The current ebuilds make
it look as if this is the case.  With shared firmware though, the build
targets are actually decoupled from the model, so renaming some of the
functions and local variable usage to indicate as such.

BUG=b:65751188
TEST=FEATURES=test emerge-coral chromeos-base/chromeos-firmware-coral

Change-Id: Ic7f8b66ec723e34d4c5fe76425a765c09769d370
Reviewed-on: https://chromium-review.googlesource.com/671636
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>