hwid.eclass: add IUSE line
Since we're checking this USE flag, make sure the eclass declares it.
The goal of the usage is to make sure the board is opting in (i.e. it
has USE=hwid_override set in its profiles), not to check if the ebuild
has it listed in its IUSE.

BUG=None
TEST=precq passes

Change-Id: Ie66d6443762f74dbb308767af06f90bd21072646
Reviewed-on: https://chromium-review.googlesource.com/682938
Commit-Ready: Edward Jee <edjee@google.com>
Tested-by: Edward Jee <edjee@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Edward Jee <edjee@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>