cros-kernel2: Unset CFLAGS
In general the kernel doesn't expect CFLAGS to be set.  As described
in the text of this change:

The kernel doesn't use CFLAGS and doesn't expect it to be passed in.
Let's be explicit that it won't do anything by unsetting CFLAGS.

In general the kernel manages its own tools flags and doesn't expect
someone external to pass flags in unless those flags have been very
specifically tailored to interact well with the kernel Makefiles.  In
that case we pass in flags with KCFLAGS which is documented to be not
a full set of flags but as "additional" flags. In general the kernel
Makefiles carefully adjust their flags in various sub-directories to
get the needed result.  The kernel has CONFIG_ options for adjusting
compiler flags and self-adjusts itself depending on whether it detects
clang or not.

NOTE: Right now, due to what's probably a bug, having CFLAGS set when
you compile the kernel actually _does_ affect something.  If you have
the right set of options then the kernel's Makefile will build stuff
from the tools/ Makefile.  In that particular case the kernel actually
uses CFLAGS for building host tools.  We should probably fix that in
the kernel itself so that when it calls into tools/ it passes
HOSTCFLAGS (which the main kernel Makefile uses) in as CFLAGS (which
the tools Makefiles use).

BUG=chromium:767565
TEST=Compile a bunch of kernels

Change-Id: If695dbd0b3eb9015bf4806de8818c7512e649026
Reviewed-on: https://chromium-review.googlesource.com/696185
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>