cros-workon: use --safe-links when copying local source
While we allow (or at least don't disallow) relative/local symlinks,
symlinks to pointing outside of the local tree are of dubious value.
In the case of cros_workon_make, it leads to bad interactions: we
first create a link in the source repo pointing outside of it, then
building via emerge retrains that symlink and the build writes under
it leading to sandbox failures.

BUG=None
TEST=build still works

Change-Id: I175c721fbf001e14b8353427f90a865733639253
Reviewed-on: https://chromium-review.googlesource.com/700723
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>