libbrillo: drop bootstat dep
The bootstat code used to be part of libbrillo directly.  When it was
split out in CL:289240, a dep on bootstat was added to libbrillo to
maintain the dep tree.  But at this point, anything that uses bootstat
should be depending on it directly (and a spot check of cs/ shows that
is the case), so drop it from libbrillo so bootstat itself can depend
on libbrillo :).

BUG=None
TEST=precq passes

Change-Id: I98025202c7f4beb745436694adfb8a388e6c0aff
Reviewed-on: https://chromium-review.googlesource.com/703535
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>