compiler-rt: skip install stage for armv7a-cros-linux-gnueabihf
We want compiler-rt for armv7a-cros-linux-gnueabihf toolchain. But
it conflicts with the one from armv7a-cros-linux-gnueabi. As a
workaround, we do not install it for armv7a-cros-linux-gnueabihf but
we keep the dependency. To use the armv7a-cros-linux-gnueabihf
toolchain, we need the board include armv7a-cros-linux-gnueabi toolchain
as well.

BUG=chromium:711369
TEST=cbuildbot chromiumos-sdk with armv7a-cros-linux-gnueabihf added
     to a board with armv7a-cros-linux-gnueabi

Change-Id: I495fe0bd286e802f9c859d99be15cb9ed0998f68
Reviewed-on: https://chromium-review.googlesource.com/710859
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>