cros-kernel2: Don't strip local symbols from kernel modules
splitdebug will normally use `strip --strip-unneeded` which deletes
local (static) symbols.  This severely affects the kernel's ability to
print useful backtraces.

The problems affect arm64 and x86_64 the worst since on arm32 the
compiler / linker tend to keep quite a few local symbols for their own
purposes.

Depending on the module, the lack of local symbols can either make the
stack crawl _very_ hard to interpret (due to the kernel just dumping a
raw hex number into the backtrace) or sorta hard to interpret (due to
the kernel thinking we're part of whichever closest function it has a
symbol for).

Adding the symbol information adds some overhead both in terms of disk
space used and in terms of RAM used.  For disk space used on a few
representative kernels, we've got:
  arm32 (kernel 3.14):  7000K =>  7412K
  arm64 (kernel 4.4):  11696K => 13344K
  x86_64 (kernel 4.4): 17472K => 19148K
...AKA, it's not huge and we can save that space elsewhere it need be
(in fact, CL:729700 ought to do the trick).  In terms of RAM used,
there will be a little bit but it should be _less than_ the above
because we don't load all the modules into memory at once.  I could
also believe (but didn't confirm) that the kernel stores these strings
more efficiently than the elf format.

In order to get this behavior that is different than the portage
default, we've got to work behind portage's back a little bit.
Specifically:
a) We've got to tell portage not to strip our modules.  We use the
   STRIP_MASK which seems to work well.
b) If we want to keep debug symbols ourselves (for kgdb) then we've
   got to generate those ourselves.  We'll solve that by just calling
   "modules_install" twice on the kernel, once asking the kernel to to
   the stripping (it will keep local symbols) and once asking it to
   give full debug info.

NOTE: there are a few side effects of the above change:
1. We no longer have "splitdebug" for the stuff that we stick in
   /usr/lib/debug.  We put full elf files there.  This is maginally
   (10%) bigger but it should be OK.  These still seem to work fine
   for kgdb.
2. We'll get a few bits of extra stuff in the "/usr/lib/debug" because
   modules_install adds some extra cruft, like the a few firmware
   files.  Again, it shouldn't hurt.
3. Build times will be slightly affected by this change, but how much
   depends on a lot of factors.  Historically each call into the
   kernel build system was quite slow due to lots of calls to the
   compiler and the overhead of our compiler wrappers.  That's getting
   fixed, though (see chromium:767073), at least for newer kernels.
   We're also speeding up our invocations of the compiler.  As far as
   I could tell this change was either neutral or slightly faster on
   kernel 4.4 w/ the fixes to chromium:767073.  Even on clang (the
   worst case) on kernel 4.4 the overhead was at most a couple of
   seconds even without those fixes.

BUG=chromium:775257
TEST=Check size of modules; do a stack crawl from in a module

Change-Id: Ie04c8a3c43640450145f2ff3da4bd4c8a159961f
Reviewed-on: https://chromium-review.googlesource.com/724168
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>