cros-workon: restore dest dir creation
In the previous commit here, the mkdir was dropped because I thought
rsync would create the destination dirs on demand.  While that's true
for the final dir, parent dirs are not created automatically.  Re-add
the mkdir call in case the ebuild checks things out to subdirs that
don't yet exist.

BUG=chromium:775787
TEST=`emerge-gale ap-daemons` works again

Change-Id: I710d20c4115cd47091c4e57f3c11893d4253fd90
Reviewed-on: https://chromium-review.googlesource.com/724943
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>