cros-workon: skip copying of .git dirs
The use of --safe-links when rsyncing the repos over breaks the .git
dir usage, so skip the .git repos themselves.  This is also what we
used to do before CL:700724 (by skipping all dot files), so nothing
should be relying on this path already.

BUG=chromium:776174
TEST=precq passes
TEST=emerging cros-workon packages no longer include .git/

Change-Id: I7bdfb246a53297df2abc251a3cf2980958d62539
Reviewed-on: https://chromium-review.googlesource.com/727542
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>