sys-kernel: Update kernel version scripts
Append _p1 to the kernel version if the version script runs from
a board overlay directory. This ensures that the generated kernel
version takes precedence over kernel versions generated by scripts
in chromiumos-overlay.

With this change, the version scripts can be used in both places
and don't have to be modified to run from board overlay directories.

BUG=chromium:731285
TEST=Run version script from board or private overlays

Change-Id: I8b669f51ab6b98ce6f5fd12138a553a6ac0befd3
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/737544
Reviewed-by: Mike Frysinger <vapier@chromium.org>