chromeos-ec: Simplify the ebuild for unibuild
This is currently quite complicated because we cannot rely on the master
configuration to tell us what to do. We're going to fix that, which allows
the ebuild to be simplified.

Update the ebuild to use cros_config_host_py and do exactly what it is
told.

BUG=chromium:777726
TEST=emerge-reef-uni chromeos-ec
emerge-coral chromeos-ec
See that it builds the expected targets
Change-Id: I8b87df8a3cfe095b3f16084479646e9eec730633
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/733090
Reviewed-by: Randall Spangler <rspangler@chromium.org>