dev-lang/rust: use LLVM bundled with rustc source package
The primary benefit of this change is decoupling the toolchain team's
rolling LLVM changes with the version rustc is meant to build against,
essentailly the latest stable as of that release of rustc. By using the
LLVM bundled with the rustc source, rustc is guaranteed to be compatible
with it.

The build takes about 5 minutes longer but the size of the binary
package is smaller by about 2%.

sys-llvm: ~390 MiB size, ~19 minutes build time
rust-llvm: ~385 MiB size, ~14.5 minutes build time

TEST=sudo emerge rust
BUG=None

Change-Id: Iae6f0782be0352b4a7899eb40de864e7e8ac8ed1
Reviewed-on: https://chromium-review.googlesource.com/734163
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>