Add udev rule: Limit the scheduler queue size for eMMC devices
By default, the scheduler queue size - nr_requets - is set to 128.
Given we are not supporing eMMC queueing feature, the IO device queue is 2,
leaving 126 for merging. When the eMMC storage device is saturated with
command, IO can be stuck up to 4 seconds in the queue.
Instead, we should propagate the saturation and put the processes to
sleep earlier.

Limiting the queue size to 32 allow passing
android.filesystem.cts.AlmostFullTest#testRandomRead more often.

BUG=b:65528573
TEST=test_that --iterations=10 100.127.1.97
cheets_CTS_N.x86.CtsFileSystemTestCases passes 6/10 times instead of always
failing on Caroline.

Change-Id: I819385c197d5b9a94585ef59748fd4ad42c5ca1e
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/738313