common-assets: Include Speech Synthesis only when necessary.
While the capability to synthesize speech from text is useful
in many commerical Chromebooks, there are some devices for which
it is unnecessary. It consumes over 100 MB of storage, which
can be a problem on resource-constrained devices.
By making this capability conditional on the "tts" USE flag,
we allow individual boards to disable the flag to save space.

BUG=b:35568451
TEST=Built Chrome OS image and deployed to a real device.
Verified that speech synthesis remains be default.
Then disabled "tts" in an overlay and verified that
this functionality is removed without affecting any others.

Change-Id: I601c6694b32a566b10f95ca39c98d28a566ad13a
Reviewed-on: https://chromium-review.googlesource.com/744642
Commit-Ready: Niranjan Kumar <kumarniranjan@chromium.org>
Tested-by: Niranjan Kumar <kumarniranjan@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>