coreboot: Simplify the ebuild for unibuild
This is currently quite complicated because we cannot rely on the master
configuration to tell us what to do. We're going to fix that, which allows
the ebuild to be simplified.

Update the ebuild to use cros_config_host_py and do exactly what it is
told.

BUG=chromium:777726
TEST=emerge-reef-uni coreboot chromeos-bootimage
See that it builds the expected target

Change-Id: I450a8632ac462d33de282936aa98fd8d357293f5
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/742380
Reviewed-by: Jason Clinton <jclinton@chromium.org>