cros-firmware: Pass --legacy to pack_firmware for non-unibuild
At present the firmware packer uses the presense of --models to detect
that it should create a unibuild firmware update. We want to stop passing
this list since it is a pain to create it in the ebuild, and we want the
packer to always pack all firmware.

As a starting point, pass this flag to pack_firmware when building a
non-unbuild firmware update.

BUG=chromium:775220
TEST=emerge-reef-uni --nodeps chromeos-firmware-reef
emerge-reef --nodeps chromeos-firmware-reef

Change-Id: I7d3d4e06a3f3805cd19b0195ff9daa94d3f66659
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/743181
Reviewed-by: Jason Clinton <jclinton@chromium.org>