chromeos-cr50-scripts: A script to reset cr50
This script is a wrapper around usb_updater. It creates and displays a
qrcode from the challenge string returned by usb_updater. The cr50
is reset when a valid authorization code is entered.

BUG=b:37952913
TEST=manual
Booted factory install shim and initiated a cr50 reset. The QRCode was
generated and displayed.

Change-Id: I528040432fd032b1a1612764e45e33f1ebae16c8
Reviewed-on: https://chromium-review.googlesource.com/786090
Commit-Ready: Sam Hurst <shurst@google.com>
Tested-by: Sam Hurst <shurst@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>