sys-apps/mosys: Only install mosys_s if requested
mosys Makefile always compiles the static version of mosys as mosys_s.
Because of the way this ebuild was writen, it assumed that this target
wasn't built until LDFLAGS was modified to have -static; this is false.
Instead, the nonfatal dobin was causing the static mosys_s to be
installed alongside mosys on every board costing us 1.3MB of space on
the rootfs.

If folks want to use the static version, mosys_s, they should copy that
after doing USE=static.

BUG=chromium:788213
TEST=FEATURES=test USE=static emerge-reef mosys

Change-Id: Ie56b3f3273e2576c001f76c6cd4c1a0766540a8c
Reviewed-on: https://chromium-review.googlesource.com/788179
Commit-Ready: Jason Clinton <jclinton@chromium.org>
Tested-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>