cros-unibuild: Add a function to obtain BCS URIs for BSP ebuilds
With unified builds the information needed by BSP ebuilds is specified in
the configuration. In particular it provides a list of files that must be
downloaded for the ebuild to operate. The ebuild thus needs to set up the
SRC_URI variable.

Add a eclass function which does this for the ebuild.

BUG=chromium:779438
TEST=manually test with reef

Change-Id: I22ebc9dcda4bfd6c5f9795a34f7be373f126ec75
Reviewed-on: https://chromium-review.googlesource.com/828021
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>