sys-kernel: Only generate valid kernel versions
When generating a kernel version, the chromeos-version scripts
match any tag which starts with 'v[234].'. This is not really
what we want; only upstream version tags should be returned.
On top of that, a glob is used for both "git describe" and "egrep",
even though only "git describe" takes a glob as parameter and "egrep"
expects a regex pattern.

Provide a more specific glob as well as a regex pattern and use
accordingly. This ensures that the script returns a version number
which is valid for use as ebuild version.

BUG=chromium:653706
TEST=Build with all supported kernels

Change-Id: I5a2881da4bb2c9dd6b5040bb5c68edf0cd760b11
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/794858
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>