[llvm] Checkout and build clang-tidy.
This is part of the work to look into setting up regular static
analyzer builds in ChromeOS.  This CL updates the LLVM ebuild to
check out, build, and install clang-tidy as part of the LLVM
build.  It also updates the compiler wrapper scripts to
invoke clang-tidy (in addition to doing the regular compilation)
if the environment variable 'WITH_TIDY' is set to 1.

BUG=chromium:808076
TEST=Successfully built llvm and clang-tidy.  Successfully
emerged Chrome, using WITH_TIDY=1, and got clang-tidy warnings
in the build log.

Change-Id: I03f8ee2be87ea688b825cf36a12ba4627b256709
Reviewed-on: https://chromium-review.googlesource.com/802174
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>