chromeos-chrome: move thinlto tuning flag to EBUILD only.
We want to provide simple chrome user a switch to enable/disable
ThinLTO by modifying the gn files. We had set some flag that can
be used only when ThinLTO is enabled, if users want to disable
thinLTO, they need to modify the CFLAGS and so on. This CL moves
these flags to EBUILD only so that simple chrome workflow user
does not need to take care of this part.

BUG=chromium:707803
TEST=USE="thinlto" emerge-falco chromeos-chrome still uses these
     flag.
     Simplechrome workflow does not include these flags.

Change-Id: I8965126a86474c3a19f3080f7e032d833b0b80d1
Reviewed-on: https://chromium-review.googlesource.com/809237
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>