power_manager: Only enable Intel HDA powersave mode on devices with a battery
With CL:565448 audio power save mode is enabled statically at boot time,
independently of the power source. This exposes an issue on some
Chromeboxes with unstable output on the ALC283 audio when idle, which
results in a popping or humming noise on external speakers.

There is no USE flag to specifically identify Chromeboxes or devices with
ALC283 audio, as an approximation skip Intel HDA powersave mode on all
devices that are always AC powered.

BUG=chromium:775486, b:69234606
TEST=emerge-mccloud power_manager
  ls /build/pyro/lib/udev/rules.d/98-powerknobs-intel-hda.rules
    => does not exist
  emerge-pyro power_manager
  ls /build/pyro/lib/udev/rules.d/98-powerknobs-intel-hda.rules
    => exists
CQ-DEPEND=CL:827167

Change-Id: I5b94d235cb7009dd59ffea79cce0c0b26ea15a5f
Reviewed-on: https://chromium-review.googlesource.com/826568
Commit-Ready: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>