cros-unibuild: Tidy up function names and comments
Various things are wrong with the comments in this file. Also some of the
names are too generic. Tidy this up.

Note that --all-models is not needed with cros_config_host commands as
they generally operate on all models.

BUG=chromium:774498
TEST=emerge-coral --nodeps chromeos-config-bsp chromeos-config \
	chromeos-bsp-coral-private chromeos-firmware-coral
Check firmware output manually and see that BSP files are installed

Change-Id: If482c4176621d25fcba8cb94dac8837838b02aad
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/831146
Reviewed-by: Mike Frysinger <vapier@chromium.org>