cros-firmware.eclass: Add feature to get URI's from files/srcuris
This allows a firmware ebuild to have a file "files/srcuris" that
declares the SRC_URI's used for this ebuild. UniBuild configuration is
used to generate this file (see next CL for example). Like this:

  export BOARD=reef
  cros_workon --board=${BOARD?} start chromeos-config-bsp \
    chromeos-firmware-${BOARD?}
  emerge-${BOARD?} chromeos-config-bsp chromeos-config
  cros_config_host -c \
    /build/${BOARD?}/usr/share/chromeos-config/config.dtb \
    get-firmware-uris > files/srcuris
  ebuild-${BOARD?} chromeos-firmware-${BOARD?}-9999.ebuild manifest
  emerge-${BOARD?} chromeos-firmware-${BOARD?}

This should solve the CQ uprev problem because there is now a file
changed within the ebuild (as opposed to outside of it).

BUG=chromium:794694
TEST=followed instructions above (with modified reef ebuild in next CL)

Change-Id: Ia41dc9cbbe3ae866de9719d292cbca324c2fe042
Reviewed-on: https://chromium-review.googlesource.com/831236
Commit-Ready: Jason Clinton <jclinton@chromium.org>
Tested-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>