infineon-firmware-updater: Break out availability check.
This change separates the code that examines TPM state and available
firmware update binaries on the root file system into a separate
script. This allows to run the check whether an update is available to
run later in the boot process in preparation to fix a boot time
regression.

BUG=chromium:780817, chromium:768342
CQ-DEPEND=CL:842406,CL:824608,CL:822339
TEST=Manual

Change-Id: I9ba4602afcc7051659214e60c014cc84e8e0b75d
Reviewed-on: https://chromium-review.googlesource.com/842724
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>