chromeos-chrome: remove webcore debug symbols on arm.
The size limit of elf32 is 4GB, and the size of .dwp has exceeded this
limit. As a result, the .dwp is corrupt and we could not get any symbols
from it. This CL removes the debug symbols of webcore as a workaround
and we need to fix this in other ways.

BUG=chromium:792999
TEST=the new .dwp file is 2.8 GB.

Change-Id: I567e0fd90ac806630d2f6e1e7281cd2ef072553f
Reviewed-on: https://chromium-review.googlesource.com/845116
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>