arc-camera3: build 3A sandbox service by default
Build 3A sandbox service by default since at the moment all projects with
arc-camera3 have 3A sandbox enabled. In the future, if an arc-camera3
project does not need 3A sandbox or is under development, we can either
disable it or leave it on since the 3A sandbox service does not
interfere with the HAL if it is not used.

BUG=b:71039230
TEST=emerge-${BOARD} arc-camera3

Change-Id: I64b35751c7475517f5e6a2a7a08c236800b4190b
Reviewed-on: https://chromium-review.googlesource.com/846640
Commit-Ready: Hung-yu Wu <hywu@chromium.org>
Tested-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>